feat(products-by-id): add lambda function
