+++ product-service/src/functions/getProductsById/mock.json
@@ -1,0 +1,6 @@
+{
+  "headers": {
+    "Content-Type": "application/json"
+  },
+  "body": "{\"name\": \"Frederic\"}"
+}
