+++ product-service/src/functions/getProductsById/schema.ts
@@ -1,0 +1,7 @@
+export default {
+  type: "object",
+  properties: {
+    name: { type: 'string' }
+  },
+  required: ['name']
+} as const;
