+++ product-service/src/functions/getProductsList/handler.ts
@@ -1,16 +1,9 @@
-import type { ValidatedEventAPIGatewayProxyEvent } from "@libs/api-gateway";
 import { formatJSONResponse } from "@libs/api-gateway";
 import { middyfy } from "@libs/lambda";
-import { PRODUCTS_MOCK } from "./mock";
+import { PRODUCTS_MOCK } from "@mocks/products";
 
-import schema from "./schema";
-
-const getProductList: ValidatedEventAPIGatewayProxyEvent<
-  typeof schema
-> = async () => {
-  return formatJSONResponse({
-    products: PRODUCTS_MOCK,
-  });
-};
+const getProductList = async () => formatJSONResponse({
+  products: PRODUCTS_MOCK,
+});
 
 export const main = middyfy(getProductList);
