--- product-service/src/functions/getProductsList/mock.ts
@@ -1,35 +1,0 @@
-type Product = {
-  id: number;
-  title: string;
-  description: string;
-  price: number;
-};
-
-export const PRODUCTS_MOCK: Product[] = [
-  {
-    id: 1,
-    title: "Reg",
-    description: "From Made in Abyss, a figure of Reg is here.",
-    price: 1000,
-  },  {
-    id: 2,
-    title: "Nanachi",
-    description: "From Made in Abyss, a figure of Nanachi is here.",
-    price: 1000,
-  },  {
-    id: 3,
-    title: "Mitty",
-    description: "From Made in Abyss, a figure of Mitty is here.",
-    price: 1000,
-  },  {
-    id: 4,
-    title: "Riko",
-    description: "From Made in Abyss, a figure of Riko is here.",
-    price: 1000,
-  },  {
-    id: 5,
-    title: "Prushka",
-    description: "From Made in Abyss, a figure of Prushka is here.",
-    price: 1000,
-  },
-];
