+++ product-service/src/functions/index.ts
@@ -1,1 +1,2 @@
 export { default as getProductsList } from './getProductsList';
+export { default as getProductsById } from './getProductsById';
