+++ product-service/tsconfig.paths.json
@@ -3,7 +3,8 @@
     "baseUrl": ".",
     "paths": {
       "@functions/*": ["src/functions/*"],
-      "@libs/*": ["src/libs/*"]
+      "@libs/*": ["src/libs/*"],
+      "@mocks/*": ["src/mocks/*"]
     }
   }
 }
